Initial support for /account/verify API.
Deca added another step in the login flow: /account/verify

It takes your GUID, password, and HWID, and gives back an accessToken.
The accessToken is then used in place of a password when calling other
API endpoints.

The accessToken is invalidated every 15 days, or any time you log in.
Instead of assuming it'll always be 15 days, it also returns
AccessTokenExpiration, which is not currently being used.

The error messages are different than they used to be, so they are not
being distinguished yet.